requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Additional dependencies for DeepSeek BDD Evaluator
openai>=1.12.0              # OpenAI client (compatible with DeepSeek API)
anthropic>=0.18.0           # Anthropic Claude API client
requests>=2.31.0            # HTTP requests
openpyxl>=3.1.0            # Excel file handling
xlsxwriter>=3.1.0          # Excel writing with formatting

# Statistical analysis and evaluation metrics
scikit-learn>=1.3.0        # Machine learning utilities
rouge-score>=0.1.2         # ROUGE metrics
nltk>=3.8.0               # Natural language toolkit
bert-score>=0.3.13        # BERTScore metric
sentence-transformers>=2.2.0  # Sentence embeddings

# Utility libraries
pyyaml>=6.0               # YAML configuration files
python-dotenv>=1.0.0      # Environment variable management
tqdm>=4.65.0              # Progress bars
backoff>=2.2.1            # Retry with exponential backoff
rich>=13.0.0              # Beautiful terminal output

# Logging
loguru>=0.7.0             # Advanced logging (optional)

# Development and testing (optional)
pytest>=7.0.0             # Testing framework
black>=23.0.0             # Code formatting
isort>=5.12.0             # Import sorting
jupyter>=1.0.0            # Notebook environment